Restructure adaptive-expressions tests
This commit separates discrete logical test cases into separate
`describe`/`it` groups using the conventional mocha axioms.

Fixes #2605
